Support Python wheels larger than 10mb
